backend/requirements.txt
# API and web server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
asyncpg==0.28.0

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
aiofiles==23.2.1
requests==2.31.0
python-slugify==8.0.1

# Smart contract analysis
web3==6.11.1

# Testing
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.0

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.0
